backend/requirements.txt
Flask==2.3.3
flask-cors==4.0.0
ultralytics==8.0.196
numpy==1.24.3
opencv-python-headless==4.8.1.78
pillow==10.0.1
torch==2.0.1
torchvision==0.15.2
requests==2.31.0
